lambda/py/requirements.txt
ask-sdk-core>=1.10.2
python-i18n>=0.3.0
